requirements.txt
Django==3.1
psycopg2-binary==2.8.6
Pillow==8.1
djangorestframework
python-decouple
pytils
django-filter
drf-yasg